--- example/server-prisma/src/trpc/source.ts
@@ -1,27 +1,0 @@
-import type { ExecutionPlanNode, SourceOrder } from '@nkzw/fate/server';
-
-const toPrismaOrderBy = (orderBy: SourceOrder) =>
-  orderBy.map((entry) => ({ [entry.field]: entry.direction }));
-
-export const prismaConnectionArgs = ({
-  cursor,
-  direction,
-  node,
-  skip,
-  take,
-}: {
-  cursor?: string;
-  direction: 'backward' | 'forward';
-  node: ExecutionPlanNode<any, any>;
-  skip?: number;
-  take: number;
-}) => ({
-  ...(cursor
-    ? {
-        cursor: { id: cursor },
-        skip,
-      }
-    : null),
-  orderBy: toPrismaOrderBy(node.orderBy),
-  take: direction === 'forward' ? take : -take,
-});
